+++ src/test/java/org/prgrms/memory/VoucherDBMemoryTest.java
@@ -1,96 +1,31 @@
 package org.prgrms.memory;
 
-import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
-import static com.wix.mysql.ScriptResolver.classPathScript;
-import static com.wix.mysql.config.Charset.UTF8;
-import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
-import static com.wix.mysql.distribution.Version.v8_0_11;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.wix.mysql.EmbeddedMysql;
-import com.zaxxer.hikari.HikariDataSource;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import javax.sql.DataSource;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.prgrms.voucher.discountType.DiscountAmount;
 import org.prgrms.voucher.discountType.DiscountRate;
 import org.prgrms.voucher.voucherType.FixedAmountVoucher;
 import org.prgrms.voucher.voucherType.PercentDiscountVoucher;
 import org.prgrms.voucher.voucherType.Voucher;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.dao.DuplicateKeyException;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 @SpringJUnitConfig
 @ActiveProfiles("jdbc")
-@TestInstance(Lifecycle.PER_CLASS)
-public class VoucherDBMemoryTest {
-
-  @Configuration
-  @ComponentScan(basePackages = {"org.prgrms.memory"})
-  static class Config {
-
-    @Bean
-    public DataSource dataSource() {
-      return DataSourceBuilder.create()
-          .url("jdbc:mysql://localhost:10001/voucher_app")
-          .username("test")
-          .password("test1234!")
-          .type(HikariDataSource.class)
-          .build();
-    }
-
-    @Bean
-    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-      return new JdbcTemplate(dataSource);
-    }
-
-  }
-
+public class VoucherDBMemoryTest extends JdbcBase{
 
   @Autowired
   VoucherDBMemory memory;
-
-  @Autowired
-  DataSource dataSource;
-
-  EmbeddedMysql embeddedMysql;
-
-  @BeforeAll
-  void setup() {
-    var mysqlConfig = aMysqldConfig(v8_0_11)
-        .withCharset(UTF8)
-        .withPort(10001)
-        .withUser("test", "test1234!")
-        .withTimeZone("Asia/Seoul")
-        .build();
-
-    embeddedMysql = anEmbeddedMysql(mysqlConfig)
-        .addSchema("voucher_app", classPathScript("schema.sql"))
-        .start();
-  }
-
-  @AfterAll
-  void cleanup() {
-    embeddedMysql.stop();
-  }
-
 
   @BeforeEach
   void clear() {
